jooq regenerate

+++ src/main/java/com/epam/ta/reportportal/jooq/Sequences.java
@@ -116,7 +116,7 @@ public class Sequences {
     /**
      * The sequence <code>public.item_attribute_id_seq</code>
      */
-    public static final Sequence<Integer> ITEM_ATTRIBUTE_ID_SEQ = new SequenceImpl<Integer>("item_attribute_id_seq", JPublic.PUBLIC, org.jooq.impl.SQLDataType.INTEGER.nullable(false));
+    public static final Sequence<Long> ITEM_ATTRIBUTE_ID_SEQ = new SequenceImpl<Long>("item_attribute_id_seq", JPublic.PUBLIC, org.jooq.impl.SQLDataType.BIGINT.nullable(false));
 
     /**
      * The sequence <code>public.launch_id_seq</code>
